test: extract VFP asset test support
